Add t/05-messages test directory
I regularly come across some tickets where we need a test (to ensure
we fixed the problem), yet that test should not be part of the Perl 6
Specification because that would control definition of what exact
wording of a message should be, for example.

So I started this directory in Rakudo's test suite for such tests.

N.B.: I realize this isn't being run by make test yet, will fix soon.