Add warning for useless use of sequence
The sequence operator (...) is not really a pure operator for purposes of
constant folding, so it's not marked pure, but we're going to carp about
using a sequence only for its side effects anyway, since it catches common
errors of the form:

    my $odds = 3, 5 ... 15;

Since this might not be detected until final sinking of the mainline, we
have to delay worrying about %*WORRIES a bit longer.  We also need to reset
the highwater so any worries produced by sinking are located at the correct
line rather than the end of the file.