Fix Range.rand generating value equal to excluded endpoints
Fixes RT#129104: https://rt.perl.org/Ticket/Display.html?id=129104

The discussion in the ticket speculates this only affects Num ranges,
however from the code I see other types are affected as well. For example
using a Rat 1.00000000000001 has the same issue in lizmat's one liner.
Reproducing with an Int is a bit harder due to much larger pool of
possible random numbers, but sometime... somewhere... in a nuclear plant
running on Perl 6, a .rand on Int range with endpoint exclusions will
produce a value equal to the excluded endpoint.

The problem is due to us checking generated value against zero and
$!min-$!max, fix by adding the $!min before doing exclusion point
equality checks. This way, the floating point noise does not introduce
a difference between what we checked against and what we return.

I measured performance impact of this on 100,000 iterations and found
it to be only 1.1% speed loss.