Fix infix:<U+2212> and prefix:<U+2212>
Currently, U+2212 infix op exists solely as a numerical op, but we
use it for stuff like Ranges and Dates too. For Ranges, it treats
the Range as a Cool. For Dates multi resolution fails.

Fix by defining a generic op that just forwards everything to texas version.
Also do the same to prefix:<U+2212> to preemptively avoid future bugs.
Currently it's defined for a single arg, but we teach it to just pass along
all the args for prefix:<->