Make sub MAIN usage ignore *% param
The code assumes a param always has a name and when it tries to get
the name of *% param, it crashes due to attempting to treat a null as
a string.

Fix by foregoing any nameless named params when printing usage.

Fixes RT#127621: https://rt.perl.org/Ticket/Display.html?id=127621