Fix Distribution::Path bin/resources file format
Currently `CURI.install: Distribution::Path.new($path)` fails to install anything in `"{$path}/resources"` and `"{$path}/bin"`. This is because `$meta<files>` expects it's keys to be relative path strings, but was ending up with absolute paths in most circumstances resulting in CURI's install method failing its check of `/^[resources|bin]\//`. This makes sure we're using the relative path string (name-path) where it should be and just absolutify the value.