Fix spurious warnings in List.reduce
When the mapper sub has associativity set, the %prec ends up having
stuff (so the first conditional passes), but the `prec` key is still
absent. So when we infix:<eq> on it, it triggers warnings about
stringifying uninitialzed values.

Fix by checking if it's inited first.

Bugfind: https://irclog.perlgeek.de/perl6/2016-12-26#i_13799375