Fix combinations() cases incorrectly returning 1-item list
Currently, asking for, say, 5-item combinations out of a 3-item list
returns a 1-item list containing an empty list. This is incorrect,
as that result applies only when we ask for 0-item combination.

Fix that by changing short-circuiting logic at the start of the
routine:

n < 1 → we have an empty list to pick from
k = 0 → can pick just 1 combination (empty list); return ((),)
n < k → we don't have enough items to pick a combination of k items; return ()

Note that negative k is not the same as k=0; it's a conscious design
decision: you can have a set of zero items, but not a set of -1 items:
https://irclog.perlgeek.de/perl6/2016-12-28#i_13810513