Fix for RT #130458
Spotted by AlexDaniel++, this patch reverts to the old behaviour
of interpreting a Uni as a .Str for regex matching.  Still not sure
that's what we want, but since we have now isolated this behaviour
into a separate Regex.ACCEPTS candidate (it was formerly taken by
the @a candidate), an adaptation of behaviour should now be simple
if so needed.