Fix infix:<cmp>(Real,Real) with 0-denominator Rationals
We have adopted[^1] IEEE 754-2008 semantics for 0-denominator rationals,
so that <0/0> is NaN, <-1/0> is -Inf, and <1/0> is Inf when .Num'ed.
However, currently infix:<cmp>(Real:D, Real:D) candidate short-cuircuits
when one of the arguments is Num(±Inf), which causes incorrect comparison
results when the other argument is one of the aforementioned Rats.

Fix by detecting whether we have a 0-denominator Rational before trying to
short-curcuit with the Num(±Inf) logic. This method is most optimal I could
find, with other methods I tried giving 50%–300% slowdowns. The current
method gives about a 4%–8% slowdown, depending on the types of Reals compared.

The worst hit of 22% was when comparing Rational against another Rational,
so I added another infix:<cmp>(Rational, Rational) candidate, which actually
made Rational cmp Rational 150% *faster*. Since this candidate is ambiguious
with Real,Real candidate, I had to stick `is default` on it, but because
`is default` uses `does` operator that doesn't appear until later in the
setting, I had to define this new infix:<cmp> candidate in src/core/operators
instead of src/core/Order.

This commit also 99% fixes the .in-range(0/0) ticket[^2], except for the
small adjustment still needed in the exception it generates.

[1] https://github.com/rakudo/rakudo/commit/1f3ca6469759c79aa809d60a7741b3a
[2] https://rt.perl.org/Ticket/Display.html?id=130452#ticket-history