Use the new setdispatcherfor op.
This corrects various bugs arising from the wrong block picking up the
dispatcher, by setting the target that it should be captured by. This
fixes a number of issues, including use of nextsame in conjunction
with multi subs with `where` clauses, and the spooky issue reported
as an OO::Monitors bug (which uses callsame) where occasionally the
callsame would not, in fact, call anything. That was due to the
finalizers being run post-GC, and the finalizer stole the dispatcher,
thus why any code that used Failure was more likely to tickle this.