Tweak dispatch:<hyper> to avoid a bug, speed up.
There's currently a code-gen bug in `"{$foo}"`, which affects multi-
threaded programs. Unfortunately, the obvious fix for it regresses a
spectest. In the meantime, we can avoid it in `>>.foo`. This is a
positive change anyway: it's cheaper to simply interpolate a variable,
and pulling the `Str` coercion up to the signature will avoid that
being done multiple times if a non-string is interpolated as the
method name. The `made` and `end-entries` methods are both called in
TAP::Harness with the `>>.foo` syntax, which led to this discovery.