Introducing Rakudo::Iterator
A dummy class for keeping (and documenting) Rakudo internal iterators.
These used to live in Rakudo::Internals, but it is getting a bit crowdy
in there.  By creating a separate class, we can also rename the iterators
to something shorter and more easily greppable.

So far these have been migrated:

  Internals::BlobbyIterator            Iterator::Blobby
  Internals::MappyIterator             Iterator::Mappy

  Internals.EmptyIterator              Iterator.Empty
  Internals.MappyIterator-values       Iterator.Mappy-values