Make endpoint-excuding string ranges work with non-alphanumeric chars
Range.iterator for the case of a Str minimum, used `.pred` and `.succ`
to adjust the minimum and maximum value for $!excludes-min or
$!excludes-max. This worked correctly for alphanumeric characters
(e.g. "a".succ -> "b"), but not for others (e.g. "!".succ -> "!").

So now $!excludes-min and $!excludes-max are passed along to the
constructor of the Iterator, so that it can adjust its start and
end code points accordingly.