Make X die on various forms of lazy sequences
- 1,2 X 1..*  now dies with:
  "Can only have single element lists before a lazy sequence in a cross"
  The reason for this is that the "2" of the first list will never be
  reached, and thus indicates a thinko of the developer.

- 1..* X 1..* now dies with:
  "Can only have one lazy sequence in a cross"
  Only the first element of the first lazy sequence will ever be used,
  and thus also indicates a thinko of the developer.

Please note that 1 X 1..* is fine, as it is then clear you only want
to cross with the one value anyway.

This should maybe be turned into a warning allowing people more rope
to hang themselves with.

This breaks one spectest for 1..* X 1..* .  I consider that test to
be faulty and giving the wrong example.