Make eqv work right on NaN and signed num zeros
eqv should return[^1] True when each side is a NaN and False
when zeros of different signs are compared. Currently, the logic always
uses `==` op, which makes it fail for those cases.

Fix by using `===` for Nums and `==` for rest of cases. We can't use
`===` for all cases, because stuff like Duration also get handled with
this op, and `===` would give False when we want a True.

[1] https://irclog.perlgeek.de/perl6-dev/2017-01-20#i_13959621
Bug find: https://irclog.perlgeek.de/perl6/2017-01-20#i_13959538