Fix infix:<cmp> for case of Real + Subclass of Rational
The current code uses .WHAT to indicate to nqp::getattr() where to
find the $!denominator attribute, which is provided by Rational role.
This works for classes that do the role directly, but not for their
subclasses, of which RatStr is one.

Fix by using public accessor instead. This is 15% slower than using
nqp to get that attribute, but works.

Fixes RT#130606: https://rt.perl.org/Ticket/Display.html?id=130606