Introducing Any.skip(n?)
Exposes the functionality of Iterator.skip-one/skip-at-least.  This
allows you to do something like:

  for "file".IO.lines.skip(10).head(10) {
  }

to get the second 10 lines of a file.  I've waited to push this until
after the 2017.01 release, so that we have about a month to revert it
should it be considered an addition we could do without  :-)