Fix infix:<xx>(..., *) for empty Slips
The current code makes the assumption that when a Slip is encountered,
it is non-empty, and it saves it in an Array and shifts from that array,
to obtain next value. This causes a Failure when the Slip is empty and
there is nothing to shift.

Fix by checking whether we can shift and returning IterationEnd if can't.

Fixes RT#130619: https://rt.perl.org/Ticket/Display.html?id=130619