Fix crash in .pull-one on IO::Handle.lines iterator
That seems to occur with RI::Batch/RI::Rotor iterators or
when .pull-one is called manually.

The iterator calls .get() on the Handle that reads from its $!PIO
and when that's exhaused the $!PIO is set to null. So when the next
.pull-one is done, we try to .get() on a closed handle by attempting
to read from nqp::null() that $!PIO is now is.

Bug find: https://irclog.perlgeek.de/perl6-dev/2017-01-27#i_13996365