Improve error reporting of `await` in 6.d.
The work so far actually made things a little worse, in so far as we
no longer called Promise.result, and thus did not get an exception
with both the original exception location and the location it was
awaited. This improves things for the new `await` implementation.
Furthermore, it means that all `await`s will get the treatment, not
just those that `await` a `Promise`. So, unlike in 6.c, `await` of a
`Supply` and a `Channel` will play ball and show both locations.