Improve `react` error reporting in 6.d.
This re-instates getting both the location of the `react` block as
well as the location of the original exception. However, the language
of the message is far better. In 6.c we report that a `Promise` was
broken, thus leaking an implementation detail of `react`. This new
erorr clearly states it was a `react` block that was at fault.