Only generate static inline info when native args.
Some years back, when Rakudo's optimizer was implemented, we didn't
run on a VM that could do inlining. This is no longer the case, and
it's much easier to do the analysis for correct semantic preservation
in a VM and get it consistently right than it is in Rakudo. A VM can
also only expend the effort on hot code. So, from now on we will only
spit out static inline info for subs with natively typed parameters.
These actually do get a notable benefit in generated code quality,
certainly beyond what MoarVM can currently achieve without the help;
JVM and JS backends will likely find it valuable too.

This also fixes various discrepancies we were getting as a result of
inlining of non-natives.