Make uniprop/unival/unimatch 2.1-0.1x as fast
Use nqp::unipropcode instead of Rakudo::Internals.PROPCODE
I may eventually remove Rakudo::Internals.PROPCODE entirely since because it
is much slower, there is not much reason for it to continue to exist.
uniprop with default arguments is 2.1x, as is uniprop-bool and uniprop-str
is about 1.9x as fast. uniprop with arguments is about 0.1-0.15x faster