Interim fix for Rational.ceiling/.floor
https://irclog.perlgeek.de/perl6/2017-02-10#i_14081234

Somehow it's possible to end up with a Rat is is <4/2>, and when
we call .ceiling/.floor, we end up with an incorrect value.

Temporarily fix by stuffing self.REDUCE-ME into these methods.
Will hunt down how to produce a <4/2> Rat and try to locate which
other place is missing the reduction that would (possibly) let us
avoid it in .ceiling/.floor