Fix crash on redeclaration check of non-sub &foo var
The redeclaration check assumes the previous thing is a Routine
and crashes when it tries to reach for $!yada that isn't present
in my &foo; Callable.

Fix by checking the symbol is a Routine before trying to check
whether it's a yada.

Bug find: https://irclog.perlgeek.de/perl6/2017-02-13#i_14093827