Revert "forbid to put variable in initializer of its declaration"
Aka “Allow self-referential arrays in declarations”

This reverts commit 3754356d8e4401c2e567f20462adb2ecd97dc7d5.

TL;DR Reverted commit does more harm than good, and with recent
changes is no longer needed as much. There are ways you can shoot
yourself in the foot, yes, and we should find a way to print awesome
error messages in these cases.

The change in the reverted commit achieves very little. In theory, it
prevented obvious ways to create a VMNull (and nulls used to blow
up). An example of the issue that was masqueraded:

    my $a := $a; say $a  # segfaulted previously

At the same time, it didn't prevent any other ways of achieving the
same thing. For example:

    my $a := $::a; say $a  # segfault, even after the fix!

However, the change prevented legitimate creation of self-referential
arrays:

    my @a = 42, @a    # ← complains needlessly

And forced to use workarounds to achieve it:

    my @a = 42, @::a  # ← does not complain and works perfectly.
                      # Congratulations, you've tricked the detection!

A proper way to fix the issue would be to prevent the segfault.
It was done in ae2ae92330e5b861b47500de0d8b0855297757a3 (actual
MoarVM/nqp commit unknown).

There's a good point that the user may have wrong assumptions on how
self-referential stuff should work with binding, and we should have a
compile-time error for this case. DIHWIDT, but would be nice if
somebody came up with a proper way to detect this.

In other words: it does not segfault anymore, so this “fix” is not
needed. And even though the behavior is LTA when you get a VMNull
anywhere, it does not warrant the prevention of a legitimate use of
self-referential arrays (or hashes, pairs, lists, etc.).

The revert had conflicts because of these commits:
* 1458f0ea3ea2943252c4436fd66a22861845f8ad
* ae651044245eabcfe172813ad3f5bab377025f51

I had to tweak it a bit to resolve the issues, should be OK.

Obviously, doesn't pass spectest because of misguided tests for the
reverted commit. This has to be corrected in errata.