Fix SC collision when identical modules are loaded from different dists
Fixes "Object does not exist in serialization context" when identical modules
are loaded from different dists which could happen when a large dist is forked
for a different auth with some modules left untouched and both forks used in
the same program.

Identical source led to identical source-id which is used by MoarVM as the
handle for identifying a serialization context body. Fix by including the
name of the installed source file in the hash. This name includes the dist-id,
so it should be unique enough.