Fix/Improve .Rat and .FatRat coercers
- Make Complex coercers take $epsilon arg since their components are Nums
- Fix Instant.Rat routing the conversion through Num;
    TAI is already a Rat, so just return it
- Implement Cool.FatRat
- Implement Numeric.FatRat that does .Rat.FatRat; any type that can provide
    a more precise FatRat should override this method.
- The Numeric and Cool additions add .FatRat coercer to these types that
    are currently missing it: Str, Complex, ComplexStr, Array, Seq, Map,
    Duration, Instant, IO::Path, Match, StrDistance

Note: no attempt has been made to sort out the coercers on :U invocants. We're
very inconsistent with teating :U numerics as `0` in many methods and IMO
that paradigm should be removed altogether.

Note 2: originally, I wanted to remove unused epsilon arg on Int and Rational
.FatRat/.Rat coercers, but there are iffy 6.c-errata tests blocking it. I've
added the removal of these params as a proposal for 6.d.