[io grant] Do not cache IO::Path.e results
Fixes RT#130889: https://rt.perl.org/Ticket/Display.html?id=130889

The problem in the ticket is due to IO::Path not setting $!e when
write happens on a write-mode opened non-existent file. However,
even if that were fixed, we're still open to issues where the file's
existence changes on the filesystem, so we can't really cache that.