Make List.reverse return a Seq
This appears to have been the intent in the speculation all along.
Reason for doing this is mainly to reduce memory pressure, as the
previous implementation would *always* create a copy of the reified
list.  By making it a Seq, it will only create a copy if that is
needed.

On a 10 element list, the gain is about 30%.  On a 1000 element list,
it's about even.  But in either case, the memory pressure is about
1/3 of the previous implementation.

This breaks some spectests because they were specifically testing for
return type.  The 1 dimensional shaped array case must still create
a copy, because otherwise .= reverse would not work correctly.

This also includes cleanup of R:In.ReverseListToList, as that is now
no longer needed/used.

Please revert if this is causing issues with the ecosystem or otherwise.