Get around X::TypeCheck::Binding in IB2
one way to make this happen is to "A-B C".split([" ", "-"]) when it tries
to swap something going via the IB2 sub. It gets a BOOTIntArray but
expects the arguments to be at least Any.

One further problem is that code like this causes a X::Binding::TypeCheck
exception that fails to generate its method due to the very lowlevel object
being used less-than-carefully.