Fix "require" with statement prefix phasers.
BEGIN require Test; failed with "Lexical with name 'Test' does not exist in
this frame". A similar error message would come for symbols in the import
list.
The reason was that the statement prefix would actually lead to a QAST::Block
being generated for the statement while the lexical stubs for the module name
and the imported symbols were created in the block containing the phased
statement. The highly specific lookup in REQUIRE_IMPORT (OUTER::CALLER::) did
not find the stubs. Fix by looking in the caller's outer lexical scope, too.