propagate purity up through meta-meta asts
We can do this because, by and large, operators must be predeclared or
the metaops wouldn't recognize them in the first place.  (The setting
violates this in spots while bootstrapping, but doesn't happen to run
afoul the "useless use" checks that depend on it.)

Previously, Zxx simply assumed that no one in their right mind would
use that with side effects on the left.  Now, since xx isn't pure,
it allows them.