Make thunk xx 42 use the new iterator
This breaks 4 tests:
  t/spec/S03-metaops/cross.t  (56-57)
  t/spec/S03-metaops/zip.t  (56-57)

I've spent several days trying to figure out why:

  my $a = 0; ($a++,) xx 1; say $a   # note the sink context

no longer gives "1".  The only reason I've been able to figure out
is that the previous non-lazy way of 'thunk xx 42' would execute
the thunks, even though the resulting Iterator would never see a
pull-one.  Now that the iterator is lazy, nobody appears to be doing
any pulling, causing the thunk not to get executed, causing $a to
remain at 0.

Further research shows that the Seq returned by infix:<xx> *is* sunk.

So, I'm not seeing why this isn't working and starting to suspect
some really basic thinko on my side, or a codegen error, or maybe
it shouldn't work at all in sink context.  In which case, the tests
should be adapted.