Better fix for RT #130979
Previous fix would just ignore any compile time errors in use arguments.
Which could lead to unexpected results in case of a typo on a use
statement for a module that *does* exist.  Not sure how we can actually
pass on the actual compile time error.  Also, there appear to be a lot
of similar cases in World.nqp that are "unprotected", which may need
attention.