Give a somewhat more helpful message
When loading a non-existing module that matches a distribution name, we found
the distribution but failed to recognize that it doesn't contain a module of
that name. We now throw a proper exception in that case.
The code is not pretty and could use some re-structuring, but at least works...