Make Parameter object available in bind error.
Just for the nominal error so far. This also allows a slight wording
improvement to the error to make clear that it was a parameter that
could not be bound. Since this is a subtype of the previous exception
type, any tests/code relying on the previous exception type will work
as before.