Log making all IO routines consistently fail()
This was part of the original IO grant proposal, but apparently was
missed in the Plan.

Currently many IO methods fail() but many IO subroutines catch those
failures and throw them. The change will remove that catching and propagate
the failures.

Per TimToady[^1], the fail was mostly invented for IO stuff, and so we'll use it

[1] https://irclog.perlgeek.de/perl6-dev/2016-12-28#i_13814044