[io grant] Fix race in &indir(IO::Path …)
Do not use IO::Path.chdir… at least for now… as it is racy. Instead,
copy over the file tests into &indir impl.

Fixes race for IO::Path argument, but not for Str, which probably
can't be fixed, since the race is inside teh coercer when Str gets
coerced into an IO::Path.