Introducing IterationSet
This commit introduces the Hash equivalent of IterationBuffer: IterationSet.
It has the same low level REPR as nqp::hash, but it will not be automatically
HLLized to a Hash when used as a parameter.  This allows usage of
nqp::hashes in common subroutines without needing to HLLize and de-HLLize
all the time.