[io grant] Restore IO::Handle.IO
After more thought, it was a bit too much to remove it along with
other IO::Pathy methods. Bringing it back, mostly so that IO() coercers
take it, just like they take Cools. And it makes far more sense to
coerce an IO::Handle to IO::Path than it does to coerce a Complex.