Implement R::Iter.FirstNThenSinkAll iterator
- Gives at most N values
- When done, calls .sink-all and calls an optional Callable
- Will be used in IO::Handle .lines, .words, .comb, and .split to auto-close
  the handle when given a limit and asked for closing

The impl isn't very DRY. I tried with a shared role for the two iterators at
first, but it first was crashing because assigning to native attributes
via a role method doesn't work. Also, I recall private methods in roles are
slow. So I duplicated a bit of stuff. The two classes really differ only in that
in one of them we avoid incrementing and counter checking on each .pull-one when
we know we don't need to.