Test error for optional params marked 'is rw'
These were copied from t/spec/S06-signature/optional.t and modified to
test the new error message, including that the name of the parameter is
in the message. The tests in t/spec/S06-signature/optional.t will be
modified to just test that an exception is thrown (and possibly its
type), not the text of the message.