[io grant] Fix up IO::Handle.Str
Ensure we coerce $!path to .Str before returning it, since it can
be a bunch of different things. Also use $.path, so that IO::Pipe
uses its `method path` instead of the :path arg given to new.

The :path attr in IO::Pipe might need more thinking; should it be
just ignored by .new and set to empty string?