[io grant] Implement IO::Path.child-secure
The original plan to make .child() resolve and check for secureness
was changed to leave .child() as is and implement .child-secure instead

There is a lot of ecosystem usage of .child(), and there's no real
path for changing it, since the proposed .concat-with() is not available
in older than bleeding-edge compilers. Another concern is the secureness
of .child would silently go away in older compilers.

Per discussion[^1], it was decided to leave .child as is, and add another
method that provides the secure version. The .concat-with will be simply
an alias to .child to make code read nicer when the .child-ed paths
aren't really children (N.B.: it was also decided to rename .concat-with
to .add)

[1] https://irclog.perlgeek.de/perl6-dev/2017-04-16#i_14435663