[io grant] Fix IO::Path.resolve with combiners; timotimo++
Stuff like "/\x[308]" is interpreted as a path separtor and a
filename named as a combiner, when we try to create files. A naive
split on '/' doesn't catch these separators however, so we implement
heavier logic where we work with bytes, swap all the dir separators
to nul bytes (that can't be in paths) and then split on nul bytes.

This also fixes the issue in IO::Path.child-secure, where the path
with a combiner is assumed to be a child, even when it isn't.