Unfatalize failure to store a repo-id in a precomp store
The repo-id is the identity (the complete state) of the whole repository
chain at the time when we resolved the dependencies for a precomp file. We
store this repo-id alongside the precomp file in the precomp store. When we
load a precomp file, we can cheaply check if this repo-id is the same as the
current repository chain's. In this case we can be sure that we can still
load the same dependencies.

There may be situations where we do not have write access to any pre-
compilation store. In this case storing the repo-id will fail, but since
that's just a cache for a little speed up it should not keep us from loading
the precomp file (we'll just have to run the dependency resolver).