Workaround for dependency resolver pulling in the Staging repo
Dependency re-resolving currently uses EVAL to unserialize a dependency
specification. This EVAL causes a reference to the serialization context
containing the Staging repository to be included in the precomp file.

The Staging repo is for when you want to precompile modules with an
unpriviledged user in a staging area that is different from where the
resulting files will be installed into. Because the unpriviledged user
won't be allowed to write to /usr/share/perl6 directly.
In other words: it's for when we create packages for distributions.
Distributions however will do their own dependency tracking. They will
compile the modules in depth first order of the dependency tree.

This means that we can be reasonably sure that the dist we are currently
packaging won't contain modules that replace one of its dependencies'
dependencies (by providing a newer version).

Thus we can skip the dependency re-resolving on detecting a change in
the repo chain (including the offending EVAL). This is strictly opt-in
by setting the RAKUDO_RERESOLVE_DEPENDENCIES environment variable to 0
when running install-dist.pl