Cope with changes to repo paths between compile and run time
When installing a module via a Staging repo, at compile time, we will find
the resources in the staging repo's path while later after installation we'll
find it in the target repo's path. Thus we need to defer creating the path
as long as possible. We do this by returning an object that does Callable
(which NativeCall will run) but also stringifies to the path for non-NativeCall
users and backwards compatibility.