Support all appropriate IO::Path methods on Distribution::Resources
%?RESOURCES<foo> used to return IO::Path objects. People started to
depend on that behavior which of course now fails as the object is no
longer an IO::Path but something that defers the creation of the path
as long as possible. Delegate calls to all of IO::Path's useful methods
(those dealing with the path name itself and for read access) to the
result of the deferred path assembly.