Fix REPL failures with Linenoise
Bug find: https://irclog.perlgeek.de/perl6/2017-04-19#i_14452772

The code for creating paths starts off by trying to `mkdir` `/` dir
and so on for each path. On OSX, this fails, despite working fine on
Linux. Also, `mkdir` used to return the list of files, but not it
returns the Failure object, 'causing the explosion.

Fix by removing all the mkpath business, because mkdir is `mkdir -p`
really. And then watch out for Failures from mkdir.