Simplify code for loading a precomp unit's dependencies
We're not supposed to know that the precompilation store deals with files
and anyway, load-unit will just return Nil if the store doesn't contain
the unit. So instead of looking for a store that has an existing file
for the dependency and then loading the unit from that store, just take
the first unit any of our stores gives us.