fix )> to work again whenever .MATCH is called
The unification of $!to and $!pos in the uncurse branch caused the )>
construct to malfunction if .MATCH was not called at the right time.  This
version uses the newly patched nqp that distinguishes $!to from $!pos again.
(Note that all internal access to $!to should be through the .to method,
since it might or might not defer to $!pos.)