Fix for RT #130760 - print perl6 usage msg to stderr when an invalid cmd line option is used
Entering an invalid option when invoking nqp or perl6 leads to an
error message being printed followed by usage instructions. In both
cases this information is written to stdout when it should have been
written to stderr. This fix solves the problem for perl6 by introducing
a flag which tells the code doing the writing which fd to use.
This fix will be dependent on nqp commit 745c61d in order to solve the
problem described above. In the absence of the nqp commit nothing will
break however, it will just continue to work the way it does today.