Speed up module loading a bit by persisting source checksum
There's no need for calculating the checksum of an installed source file over
and over, when it can change only on installation. So instead, we calculate
it at install time and record it in the short-name lookup file where we
already store the name of the source file.
With this change, we no longer read any source file when loading an installed
module (dependencies were already covered).