[io grant] Make IO::Handle.iterator a private lines iterator
Fixes unwanted automagical behaviour of for $fh { ... } looping
over the .lines from the handle instead of the handle itself.

Discussion: https://irclog.perlgeek.de/perl6-dev/2017-05-01#i_14512345

Previous two revert commits are part of the work in this commit and
are just reverting IO::ArgFiles back to using .lines instead of
no-longer-available .iterator method. It was done mostly because
`:close` arg on IO::Handle.lines was going away, but it stayewd.