[io grant] Fix IO::Handle.comb/.split; make them .slurp
The current implementation has the following issues:
- The features of the two methods do not match those of Str versions
- Incorrect results are produced when the chunk to be produced overlaps
    the $*DEFAULT-READ-ELEMS boundary; e.g. .comb(65546) will return
    a 2-element list for a file with a `"x" x 65546` string in it; one
    element will be $*DEFAULT-READ-ELEMS in size, and the other will
    be 10 chars long. The same sort of issues exist for other
    boundary-crossing cases, such as .split(\w+)

Originally, I tried to cook up a mini-version of Cat lazy string, but
my Cat-foo isn't strong enough and I hit the issue where I didn't know
how much to cut off from the start of the string.

So the solution I ended up with is to make .comb/.split .slurp the
entire file. I think when we have a proper, full-fledged Cat, we
can just use it instead and feed it via chunked reads, but for now this
will do.

[1] https://github.com/zoffixznet/perl6-Kitten