Handle QuantHash:U coercion differently
This should fix some issues with set operators (RT ##131300).

Before this, Set.Set would return Set.  Now it returns set(Set).  So
instead of not coercing a QuantHash type object, we now coerce it into
a defined QuantHash object with the type object as its member.

This is consistent with e.g. Int.Set, which already coerced to set(Int).