Re-work IO::Socket.recv to use decode stream.
This means it doesn't use VM-level string I/O any more. Doing this
also shows up that the semantics of .recv in chars mode need a more
careful look, as it seems it's possible that it could end up trying to
decode parts of multi-byte sequences (which will then blow up). The
semantics here match what we were already doing in the VM, keeping the
test suite happy. It's worth considering if there are better semantics
we could adopt that would keep it similarly happy.