Make (Bag|Mix).AT-KEY about 10% faster
By only needing to do a lookup in the underlying hash once on a match.
Keeps the original Baggy.AT-KEY implementation for subclasses: not
possible to do the (Bag|Mix) implementation in the role, as we cannot
initialize a class lexical in a role :-( .  At least not in the Setting.