[io grant] Allow no-arg &prompt
It's not entirely unheard of to not want to include any text message
to the user when prompting for some input. It's a bit antisocial to
force the user to specify `''` arg for such cases.

Worse, they might introduce a subtle bug into their code by switching
to use `get` for a message-less `prompt`, but `get` defaults to
`$*ARGFILES`, instead of `$*IN`