Eliminate nqp::printfh use in IO::Handle.
It goes away in favor of explicit calls to .encode and then calling
nqp::writefh. This is, without further optimization effort, probably
a slowdown, having glanced Str.encode. It's not enough to make any
difference to spectest time, however, so it's likely not all that
dramatic. This does regress one spectest, which seems to be trying
to do string I/O with the encoding 'bin'.